feat(limit-script): moved get subscription to payments service
