+++ src/externals/payments/payments.service.ts
@@ -3,6 +3,8 @@ import { ConfigService } from '@nestjs/config';
 import Stripe from 'stripe';
 
 import { UserAttributes } from '../../modules/user/user.attributes';
+import { HttpClient } from '../http/http.service';
+import { Sign } from 'src/middlewares/passport';
 
 @Injectable()
 export class PaymentsService {
@@ -10,7 +12,8 @@ export class PaymentsService {
 
   constructor(
     @Inject(ConfigService)
-    configService: ConfigService,
+    private configService: ConfigService,
+    private httpClient: HttpClient,
   ) {
     const stripeTest = new Stripe(process.env.STRIPE_SK_TEST, {
       apiVersion: '2023-10-16',
@@ -47,5 +50,25 @@ export class PaymentsService {
     }
 
     return false;
+  }
+
+  async getCurrentSubscription(uuid: UserAttributes['uuid']) {
+    const jwt = Sign(
+      { payload: { uuid } },
+      this.configService.get('secrets.jwt'),
+    );
+
+    const params = {
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${jwt}`,
+      },
+    };
+
+    const res = await this.httpClient.get(
+      `${this.configService.get('apis.payments.url')}/subscriptions`,
+      params,
+    );
+    return res.data;
   }
 }
